+++ consensus/merge/src/main/java/org/hyperledger/besu/consensus/merge/blockcreation/MergeCoordinator.java
@@ -15,6 +15,8 @@
 package org.hyperledger.besu.consensus.merge.blockcreation;
 
 import static org.hyperledger.besu.consensus.merge.TransitionUtils.isTerminalProofOfWorkBlock;
+import static org.hyperledger.besu.consensus.merge.blockcreation.MergeMiningCoordinator.ForkchoiceResult.Status.INVALID;
+import static org.hyperledger.besu.consensus.merge.blockcreation.MergeMiningCoordinator.ForkchoiceResult.Status.INVALID_PAYLOAD_ATTRIBUTES;
 import static org.hyperledger.besu.util.Slf4jLambdaHelper.debugLambda;
 
 import org.hyperledger.besu.consensus.merge.MergeContext;
@@ -207,15 +209,31 @@ public class MergeCoordinator implements MergeMiningCoordinator {
   }
 
   @Override
-  public Optional<BlockHeader> getOrSyncHeaderByHash(final Hash blockhash) {
+  public Optional<BlockHeader> getOrSyncHeaderByHash(final Hash blockHash) {
     final var chain = protocolContext.getBlockchain();
-    final var optHeader = chain.getBlockHeader(blockhash);
+    final var optHeader = chain.getBlockHeader(blockHash);
 
     if (optHeader.isPresent()) {
       debugLambda(LOG, "BlockHeader {} is already present", () -> optHeader.get().toLogString());
     } else {
-      debugLambda(LOG, "appending block hash {} to backward sync", blockhash::toHexString);
-      backwardSyncContext.syncBackwardsUntil(blockhash);
+      debugLambda(LOG, "appending block hash {} to backward sync", blockHash::toHexString);
+      backwardSyncContext.syncBackwardsUntil(blockHash);
+    }
+    return optHeader;
+  }
+
+  @Override
+  public Optional<BlockHeader> getOrSyncHeaderByHash(
+      final Hash blockHash, final Hash finalizedBlockHash) {
+    final var chain = protocolContext.getBlockchain();
+    final var optHeader = chain.getBlockHeader(blockHash);
+
+    if (optHeader.isPresent()) {
+      debugLambda(LOG, "BlockHeader {} is already present", () -> optHeader.get().toLogString());
+    } else {
+      debugLambda(LOG, "appending block hash {} to backward sync", blockHash::toHexString);
+      backwardSyncContext.updateHeads(blockHash, finalizedBlockHash);
+      backwardSyncContext.syncBackwardsUntil(blockHash);
     }
     return optHeader;
   }
@@ -239,10 +257,7 @@ public class MergeCoordinator implements MergeMiningCoordinator {
                 protocolContext, block, HeaderValidationMode.FULL, HeaderValidationMode.NONE);
 
     validationResult.blockProcessingOutputs.ifPresentOrElse(
-        result -> {
-          result.worldState.persist(block.getHeader());
-          chain.appendBlock(block, result.receipts);
-        },
+        result -> chain.appendBlock(block, result.receipts),
         () ->
             protocolSchedule
                 .getByBlockNumber(chain.getChainHeadBlockNumber())
@@ -254,7 +269,10 @@ public class MergeCoordinator implements MergeMiningCoordinator {
 
   @Override
   public ForkchoiceResult updateForkChoice(
-      final BlockHeader newHead, final Hash finalizedBlockHash, final Hash safeBlockHash) {
+      final BlockHeader newHead,
+      final Hash finalizedBlockHash,
+      final Hash safeBlockHash,
+      final Optional<PayloadAttributes> maybePayloadAttributes) {
     MutableBlockchain blockchain = protocolContext.getBlockchain();
     Optional<BlockHeader> currentFinalized = mergeContext.getFinalized();
     final Optional<BlockHeader> newFinalized = blockchain.getBlockHeader(finalizedBlockHash);
@@ -265,6 +283,7 @@ public class MergeCoordinator implements MergeMiningCoordinator {
         && newFinalized.isPresent()
         && !isDescendantOf(currentFinalized.get(), newFinalized.get())) {
       return ForkchoiceResult.withFailure(
+          INVALID,
           String.format(
               "new finalized block %s is not a descendant of current finalized block %s",
               finalizedBlockHash, currentFinalized.get().getBlockHash()),
@@ -284,14 +303,14 @@ public class MergeCoordinator implements MergeMiningCoordinator {
                         newHead.getBlockHash(), finalized.getBlockHash()));
 
     if (descendantError.isPresent()) {
-      return ForkchoiceResult.withFailure(descendantError.get(), latestValid);
+      return ForkchoiceResult.withFailure(INVALID, descendantError.get(), latestValid);
     }
 
     Optional<BlockHeader> parentOfNewHead = blockchain.getBlockHeader(newHead.getParentHash());
     if (parentOfNewHead.isPresent()
         && parentOfNewHead.get().getTimestamp() >= newHead.getTimestamp()) {
       return ForkchoiceResult.withFailure(
-          "new head timestamp not greater than parent", latestValid);
+          INVALID, "new head timestamp not greater than parent", latestValid);
     }
     // set the new head
     blockchain.rewindToBlock(newHead.getHash());
@@ -310,6 +329,11 @@ public class MergeCoordinator implements MergeMiningCoordinator {
               blockchain.setSafeBlock(safeBlockHash);
               mergeContext.setSafeBlock(newSafeBlock);
             });
+
+    if (maybePayloadAttributes.isPresent()
+        && !isPayloadAttributesValid(maybePayloadAttributes.get(), newHead)) {
+      return ForkchoiceResult.withFailure(INVALID_PAYLOAD_ATTRIBUTES, null, Optional.empty());
+    }
 
     return ForkchoiceResult.withResult(newFinalized, Optional.of(newHead));
   }
@@ -473,6 +497,11 @@ public class MergeCoordinator implements MergeMiningCoordinator {
     }
     // neither matching nor is the ancestor block height lower than newBlock
     return false;
+  }
+
+  private boolean isPayloadAttributesValid(
+      final PayloadAttributes payloadAttributes, final BlockHeader headBlockHeader) {
+    return payloadAttributes.getTimestamp() > headBlockHeader.getTimestamp();
   }
 
   @FunctionalInterface